Refactor test code into utility functions to avoid repetition.
