+++ tests/mocha/40-heartbeat.js
@@ -2,31 +2,15 @@
  * Copyright (c) 2024-2026 Digital Bazaar, Inc.
  */
 import {
-  listCelFiles, runDidcel, TMP_DIR
+  listCelFiles, readCelFile, runAndCapture, runDidcel
 } from './helpers.js';
 import chai from 'chai';
-import {gunzipSync} from 'node:zlib';
-import {join} from 'node:path';
-import {readFileSync} from 'node:fs';
 
 const {expect} = chai;
-
-function readCel(filename) {
-  return JSON.parse(
-    gunzipSync(readFileSync(join(TMP_DIR, 'logs', filename))).toString('utf8'));
-}
 
 const HB_COMMANDS = [
   'create', 'witness', 'heartbeat', 'witness', 'save', 'quit'
 ];
-
-async function runHeartbeat() {
-  const before = listCelFiles();
-  const result = await runDidcel({commands: HB_COMMANDS});
-  const after = listCelFiles();
-  const newFile = after.find(f => !before.includes(f));
-  return {...result, newFile};
-}
 
 describe('heartbeat', function() {
   this.timeout(120000);
@@ -43,22 +27,26 @@ describe('heartbeat', function() {
   });
 
   it('should produce a CEL with 2 events (create + heartbeat)', async () => {
-    const {exitCode, stderr, newFile} = await runHeartbeat();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: HB_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     expect(celContent).to.have.property('log');
     expect(celContent.log).to.have.length(2);
   });
 
   it('should have heartbeat event with correct operation type', async () => {
-    const {exitCode, stderr, newFile} = await runHeartbeat();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: HB_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     const heartbeatEntry = celContent.log[1];
     expect(heartbeatEntry.event.operation).to.have.property(
@@ -68,11 +56,13 @@ describe('heartbeat', function() {
 
   it('should hash-link heartbeat event to the witnessed create event',
     async () => {
-      const {exitCode, stderr, newFile} = await runHeartbeat();
+      const {exitCode, stderr, newFile} = await runAndCapture({
+        commands: HB_COMMANDS
+      });
 
       expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-      const celContent = readCel(newFile);
+      const celContent = readCelFile(newFile);
 
       const heartbeatEntry = celContent.log[1];
       expect(heartbeatEntry.event).to.have.property('previousEventHash');
@@ -80,11 +70,13 @@ describe('heartbeat', function() {
     });
 
   it('should witness the heartbeat event', async () => {
-    const {exitCode, stderr, newFile} = await runHeartbeat();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: HB_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     const heartbeatEntry = celContent.log[1];
     expect(heartbeatEntry).to.have.property('proof');
