+++ tests/mocha/50-deactivate.js
@@ -2,19 +2,11 @@
  * Copyright (c) 2024-2026 Digital Bazaar, Inc.
  */
 import {
-  listCelFiles, runDidcel, TMP_DIR
+  listCelFiles, readCelFile, runAndCapture, runDidcel
 } from './helpers.js';
 import chai from 'chai';
-import {gunzipSync} from 'node:zlib';
-import {join} from 'node:path';
-import {readFileSync} from 'node:fs';
 
 const {expect} = chai;
-
-function readCel(filename) {
-  return JSON.parse(
-    gunzipSync(readFileSync(join(TMP_DIR, 'logs', filename))).toString('utf8'));
-}
 
 const DEACTIVATE_COMMANDS = [
   'create', 'witness',
@@ -22,14 +14,6 @@ const DEACTIVATE_COMMANDS = [
   'deactivate', 'witness',
   'save', 'quit'
 ];
-
-async function runDeactivate() {
-  const before = listCelFiles();
-  const result = await runDidcel({commands: DEACTIVATE_COMMANDS});
-  const after = listCelFiles();
-  const newFile = after.find(f => !before.includes(f));
-  return {...result, newFile};
-}
 
 describe('deactivate', function() {
   this.timeout(120000);
@@ -50,22 +34,26 @@ describe('deactivate', function() {
 
   it('should produce a CEL with 3 events (create + update + deactivate)',
     async () => {
-      const {exitCode, stderr, newFile} = await runDeactivate();
+      const {exitCode, stderr, newFile} = await runAndCapture({
+        commands: DEACTIVATE_COMMANDS
+      });
 
       expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-      const celContent = readCel(newFile);
+      const celContent = readCelFile(newFile);
 
       expect(celContent).to.have.property('log');
       expect(celContent.log).to.have.length(3);
     });
 
   it('should have deactivate event with correct operation type', async () => {
-    const {exitCode, stderr, newFile} = await runDeactivate();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: DEACTIVATE_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     const deactivateEntry = celContent.log[2];
     expect(deactivateEntry.event.operation).to.have.property(
@@ -74,11 +62,13 @@ describe('deactivate', function() {
   });
 
   it('should hash-link all events in the chain', async () => {
-    const {exitCode, stderr, newFile} = await runDeactivate();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: DEACTIVATE_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     for(let i = 1; i < celContent.log.length; i++) {
       const entry = celContent.log[i];
@@ -88,11 +78,13 @@ describe('deactivate', function() {
   });
 
   it('should have witness proofs on all events', async () => {
-    const {exitCode, stderr, newFile} = await runDeactivate();
+    const {exitCode, stderr, newFile} = await runAndCapture({
+      commands: DEACTIVATE_COMMANDS
+    });
 
     expect(exitCode, `stderr: ${stderr}`).to.equal(0);
 
-    const celContent = readCel(newFile);
+    const celContent = readCelFile(newFile);
 
     for(const entry of celContent.log) {
       expect(entry).to.have.property('proof');
